Modernized CMake setup
Moved to CMake 2.8.11 requirement (it is still old, ubuntu 14.04 LTS is on
2.8.12).
Used a proper annotated target for the generated libs.
Tests are linked to the library target -> -I and -L are automatic. No
more install before make test.
List of header files removed from src/CMakeLists: it was unused and out
of date with respect to file tree.
install now creates a CMake file containing target definitions for users
of cmake (<prefix>/share/bpp/).

TODO: improve the main CMakeList:
- add proper targets to other bpp-*, and use a more standard
  find-package macro
- NO_DEP_... : is it still needed ?
- remove CTest stuff that is only useful when pushing to elsewhere ?