Improved DF graph debugging.
Added system to provide custom descriptions to nodes.
Requires that node impl (custom ones, templates are managed) provide
their own, so not perfect.
Allows DF graph for lik example to be readable.
This system may not be that much better than the automatic one for
bigger graphs, so it might be removed later.