Removed All dynamic exception specifications
Dynamic exceptions specifications (throw (Exc1, Exc2, ...)) are
deprecated in C++11.
They are replaced by the noexcept specification, which can be used to
indicate that a function can never throw exceptions.
Functions can throw exceptions by default, so all specifications have
been removed instead of using noexcept (false) which is not very useful.

The main reason for this upgrade is G++ 7, which  includes -Wdeprecated in -Wall, and generates huge amounts of warnings.