Optional demangling support without cmakery
Demangling requires cxxabi, which was checked with cmake stuff.
As it is completely optional (debug info for DF graph), use a simpler
approach.

New approach just checks for compiler, and enables demangling for gcc
and clang.
Testing with godbolt shows that C++11-enabled versions of these
compilers always have cxxabi.h and demagling support.

Fallback to no demangling if not found.