Added DF derivation semantics
Added doc for semantics

Corrected code with the correct semantics:
df/dx is the derivative of expression f with respect to expression x.
Even if x is a constant, x is considered a variable in the derivation.
Thus dx/dx == 1 event if x is constant.

-> Added checks to all numerical nodes:
if deriving with respect to self, always return a 1 constant.