Restore definitions of likelihood nodes
Most are simple typedefs to numeric nodes.
Added better documentation on what is done, why, and which types and
representations are used.

Removed specific dimension types (and their file).
Instead, constructor functions can be used in addition to the text doc.

Use RowVector for equilibrium frequencies instead of Vector.
This way MatrixProduct is sufficient for all uses and the transposed
variants are unneeded.

Added SumOfLogarithms to dataflow numeric toolbox.
This can be reused in other contexts than the likelihood example.
This is now generic in the argument type.
TotalLogLikelihood is now just a typedef to it.