Fix targetDimension problems
Added underscores to targetDimension attributes of dataflow classes for
homogeneity.

Made targetDimension a mandatory constructor argument.
Before, if ommited it would default to the Dimension<T> default value.
This was useful to simplify code instantiating nodes manipulating basic
types like double, as they do not use the dimension field, and the
default value is ok.
However this lead to an undetected bug where the default value (0,0) was
used for matrices because the targetdimension constructor argument was
forgotten.
By making the argument mandatory, this kind of bug is prevented at the
cost of brevity for code using non matrix types.
However, as dataflow instantiation should be internal, the loss of
brevity should not be a problem.