Start of small example ; fixes and cleaning
Started to convert Phylogeny.h to a simple header-only inline function
to create a likelihood example.

Noticed a bug in the use of numeric nodes : bad order of coordinates for
matrix-related computations.

Likelihoods are stored site by site, with each column being a site.
This is important for data locality.
As a consequence (see comments), transition matrix must be transposed before multiplication (TODO).

Removed most useless files to clean a bit, only Phylogeny.cpp is useful
for reference now.