Added NumericalDependencyTransform to enable transposition
We need variants of matrix product that transpose one of the matrix
before the product.
Writing them by hand would require finding a new node naming scheme to
describe whether the first or second dep is transposed.
Code unaffected by transposition would need to be duplicated many times.

To avoid that, I introduce NumericalDependencyTransform<T>.
This struct is used to implement transformations of dependency values
before a computation.
Transformations are trigerred by using type tags as dependency types,
like this: MatrixProduct<R, Transposed<T0>, T1> for R = tr(T0) * T1.
The template struct NumericalDependencyTransform implements the
transformation with specialisation for these type tags.

Transposition could be done by a separate dataflow node, but that would
be inefficient (intermediate copy of matrices).
This system allows to integrate transposition in other node computations
if needed.
The system itself can support other types of transformations, with
limitations:
- transformations cannot combine multiple dependencies
- they must not change the semantics of hasNumericalProperty()
Thus it is unlikely that many other transformations will be implemented
this way in the future.

Support of Transposition has been added to MatrixProduct for the
likelihood case, and to Convert because the code of MatrixProduct
required it.
This means that convert can be used as a standalone node performing a
transposition if needed.
Only a small test of Convert has been added.