Restored most of the likelihood example
- Use the transposed matrix multiply for Forward lik computation
- add dimension doc to all Likelihood nodes
- added code to generate most of the DF graph (WIP)
- used the code in a restored new_lik.cpp test file.

The pair of mutually recursive functions is difficult to implement
cleanly.
The biggest problem is that lambdas cannot be recursive.
The current solution is to generate a lambda like struct with two
mutually recursive methods.
This avoid using helper functions with 7 or more arguments containing
the recursion state.

This example does not interface with bpp::Derivable and bpp::Optimizer
classes, so automatic optimisation is not possible for now.
What will be possible to do is to compare likelihood values for
correctness, and manually asking for the derivatives using node->deriveAsValue(c, x);
Unfortunately, without merging the derivatives will be sub-optimal, as
many things will be duplicated.

The current code is not finished ; initial cond lik is not computed
(for next commit).